[DF-7180] warning fixed
Former-commit-id: ef1245a1f4304193530ca24246896f24982ef322

+++ Sources/Tools/QtTools/ConsoleWidget/LogFilterModel.cpp
@@ -15,7 +15,7 @@ LogFilterModel::~LogFilterModel()
 
 void LogFilterModel::SetFilters(const QVariantList & _filters)
 {
-    quint64 tmpFilters = 0;
+    int tmpFilters = 0;
     for(const auto &filter : _filters)
     {
         DVASSERT(filter.canConvert<int>());
